requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
django-debug-toolbar>=3.2.3,<3.6
django-templated-mail>=1.0.0,<1.2
requests>=2.20.0,<2.28.2
psycopg2>=2.8.6,<2.9
redis>=4.2.1,<4.4
celery>=5.2.3,<5.3
drf-spectacular>=0.15.1,<0.16
